requirements.txt
requests
django
djangorestframework
django-cors-headers
validate_email